+++ FHIR-Converter/client/src/core/common/enum/template-type.ts
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License in the project root for license information.
+ */
+
+export enum TemplateType { 
+	ccda = 'C-CDA', 
+	hl7v2 = 'HL7v2' 
+}